feat: add thread helps, refactor: rename file to singular
